port toLocaleOrdinal to LocaleContextProvider
