+++ src/components/LocaleContextProvider.tsx
@@ -50,6 +50,9 @@ type LocaleContextProps = {
     /** Gets the locale digit corresponding to a standard digit */
     toLocaleDigit: (digit: string) => string;
 
+    /** Formats a number into its localized ordinal representation */
+    toLocaleOrdinal: (number: number) => string;
+
     /** Gets the standard digit corresponding to a locale digit */
     fromLocaleDigit: (digit: string) => string;
 
@@ -65,6 +68,7 @@ const LocaleContext = createContext<LocaleContextProps>({
     updateLocale: () => '',
     formatPhoneNumber: () => '',
     toLocaleDigit: () => '',
+    toLocaleOrdinal: () => '',
     fromLocaleDigit: () => '',
     preferredLocale: CONST.LOCALES.DEFAULT,
 });
@@ -98,6 +102,8 @@ function LocaleContextProvider({preferredLocale, currentUserPersonalDetails = {}
 
     const toLocaleDigit = useMemo<LocaleContextProps['toLocaleDigit']>(() => (digit) => LocaleDigitUtils.toLocaleDigit(locale, digit), [locale]);
 
+    const toLocaleOrdinal = useMemo<LocaleContextProps['toLocaleOrdinal']>(() => (number) => LocaleDigitUtils.toLocaleOrdinal(locale, number), [locale]);
+
     const fromLocaleDigit = useMemo<LocaleContextProps['fromLocaleDigit']>(() => (localeDigit) => LocaleDigitUtils.fromLocaleDigit(locale, localeDigit), [locale]);
 
     const contextValue = useMemo<LocaleContextProps>(
@@ -109,10 +115,11 @@ function LocaleContextProvider({preferredLocale, currentUserPersonalDetails = {}
             updateLocale,
             formatPhoneNumber,
             toLocaleDigit,
+            toLocaleOrdinal,
             fromLocaleDigit,
             preferredLocale: locale,
         }),
-        [translate, numberFormat, datetimeToRelative, datetimeToCalendarTime, updateLocale, formatPhoneNumber, toLocaleDigit, fromLocaleDigit, locale],
+        [translate, numberFormat, datetimeToRelative, datetimeToCalendarTime, updateLocale, formatPhoneNumber, toLocaleDigit, toLocaleOrdinal, fromLocaleDigit, locale],
     );
 
     return <LocaleContext.Provider value={contextValue}>{children}</LocaleContext.Provider>;
